backend/requirements.txt
boto3
PyJWT
requests